Change the card link to only use the main content
The card is no longer a link itself. The main target is moved to be
the content (the area containing the title and tags). This is in
preparation of upcoming changes: if the whole card is a link, it
cannot contain other links, due to a restriction by html. Later a card
may have more links to provide.