Don't use jq to clean whitespace/quotes
It seems that older jq versions choke on some output from curl inside
these scripts. But I couldn't reproduce this on the cmdline; so not
really sure what's going on, suspecting some settings in the bash
script. Tested this with jq-1.5 and jq-1.6, both worked. But since
`xargs` also trims a string from quotes and whitespace, it can be used
instead of jq in these places.