Fix build when bloop plugin is present
When using sbt-bloop, the build doesn't compile anymore. The reason
seems to be incomptible `sbt-io` dependencies pulled in from
`sbt-bloop` and `sbt-native-packager` (as well as `sbt-github-pages`).
Interestingly, the build compiles fine if either one of these plugins
is removed. Only together with `sbt-bloop` the build fails to compile.
The workaround is to explicitely pull in the io depenency based on the
sbt version in use.